The income of municipalities of "Emilia Romagna".

The dataset contains the total income of the municipalities in the region "Emilia Romagna", in Italy, for the year 2015. Each municipality is defined by their own ISTAT (Istituto nazionale di statistica, Italy) code and a name. For each municipality there are the following auxiliary variables: province, number of taxpayers and spatial coordinates (geographical position).

income_emilia

Format

A data frame with 334 rows and 7 variables:

municipality_code

identification municipality code

municipality

name of the municipality

province

province of the municipality

numtaxpay

number of taxpayers in the municipality

tot_inc

average income of the municipality

x_coord

coordinate x of the municipality

y_coord

coordinate y of the municipality

Source

The dataset is a rearrangement from the data released by the Italian Finance Department, MEF - Dipartimento delle Finanze (Italy).