Standardize a symmetric matrix (distances) to fixed row (column) products

Standardizes a distance matrix to fixed rows and columns products. The function iteratively constrains a logarithmic transformed matrix to know products, and in order to keep the symmetry of the matrix, at each iteration performs an average with its transpose. When the known products are all equal to a constant (e.g. 0), this method provides a simple and accurate way to scale a distance matrix to a doubly stochastic matrix.

stprod(mat, vec, differ = 1e-15, niter = 1000L)

Arguments

mat
A distance matrix size NxN.

vec
A vector of row (column) constraints.

differ
A scalar with the maximum accepted difference with the constraint (default = 1e-15).

niter
An integer with the maximum number of iterations (default = 1000).

Value

Returns a standardized distance matrix of size NxN.

Details

The standardized matrix will not be affected by problems arising from units with different inclusion probabilities caused by undesired features of the spatial distribution of the population, as edge effects and/or isolated points.

References

Benedetti R, Piersimoni F (2017). A spatially balanced design with probability function proportional to the within sample distance. Biometrical Journal, 59(5), 1067-1084. https://doi.org/10.1002/bimj.201600194

Examples

# NOT RUN {
d <- matrix(runif(200), 100, 2)
dis <- as.matrix(dist(d))
con <- rep(0, nrow(dis))
stand_dist <- stprod(mat = dis, vec = con)
# }# NOT RUN {
dis <- as.matrix(dist(cbind(simul1$x, simul1$y))) # distance matrix
con <- rep(0, nrow(dis)) # vector of constraints
stand_dist <- stprod(mat = dis, vec = con) # standardized matrix
# }