Sum Within Distance (Spatially Balanced Sampling Design)

This is an implemention of a spatially balanced design, with a probability function proportional to the within sample distance, using the sum of distance as an index of the within sample distance (Sum Within Distance, swd in short). To have a constant inclusion probabilities \(\pi_{i}=nsamp/N\), where \(nsamp\) is sample size and \(N\) is population size, standardize the distance matrix with function stsum.

swd(dis, nsamp, bexp, nrepl = 1L, niter = 10L)

Arguments

dis
A distance matrix NxN that specifies how far are all the pairs of units in the population.

nsamp
Sample size.

bexp
Parameter \(\beta\) for the algorithm. The higher \(\beta\) is, the more the sample is going to be spread.

nrepl
Number of samples to draw (default = 1).

niter
Number of iterations for the algorithm. More iterations are better but require more time. Usually 10 is very efficient (default = 10).

Value

Return a matrix nrepl x nsamp, which contains the nrepl selected samples, each of them stored in a row. In particular, the i-th row contains all labels of units selected in the i-th sample.

References

Benedetti R, Piersimoni F (2017). “A spatially balanced design with probability function proportional to the within sample distance.” Biometrical Journal, 59(5), 1067–1084.

Examples

# NOT RUN {
# Example 1
# Draw 20 samples of dimension 15 without constant probabilities and beta = 1
dis <- as.matrix(dist(cbind(income_emilia$x_coord, income_emilia$y_coord))) # distance matrix
nsamp <- 15  # sample size
nrepl <- 20  # number of samples to draw
niter <- 10  # number of iterations in the algorithm
bexp <- 10   # parameter beta
samples <- swd(dis, niter, nsamp, nrepl, bexp)  # drawn samples
# }# NOT RUN {
# Example 2
# Draw 20 samples of dimension 15 with constant probabilities equal to nsamp/N and beta = 10
# with N = population size
dis <- as.matrix(dist(cbind(income_emilia$x_coord,income_emilia$y_coord))) # distance matrix
nsamp <- 15  # sample size
nrepl <- 20  # numbers of samples to drawn
niter <- 10  # numbers of iterations in the algorithm
bexp <- 10   # parameter beta
vec <- rep(1, nrow(dis)) # vector of constraints
stand_dist <- stsum(dis, vec, 1e-15, 1000) # standardized matrix
samples <- swd(stand_dist, niter, nsamp, nrepl, bexp)  # drawn samples
# }